requirements.txt
pydantic-mongo @ git+https://github.com/laruss/pydantic-mongo
python-decouple~=3.8
flask-cors~=4.0.0